Get reworked RSX to compile
