+++ rpcs3/Emu/RSX/NV47/FW/reg_context.cpp
@@ -1,0 +1,1 @@
+#include "stdafx.h"
