+++ rpcs3/Emu/RSX/NV47/FW/reg_context.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <util/types.hpp>
+
+namespace rsx
+{
+	// TODO: Basically replaces parts of the current "rsx_state" object
+	struct reg_context
+	{
+		u32 registers[1];
+	};
+}
