+++ rpcs3/Emu/RSX/NV47/HW/nv0039.h
@@ -1,0 +1,12 @@
+// NV47 Buffer Management
+#pragma once
+
+#include "context.h"
+
+namespace rsx
+{
+	namespace nv0039
+	{
+		void buffer_notify(context* ctx, u32 reg, u32 arg);
+	}
+}
