+++ rpcs3/Emu/RSX/NV47/HW/nv3089.h
@@ -1,4 +1,6 @@
+// NV47 2D Blit Engine
 #pragma once
+
 #include "context.h"
 
 namespace rsx
