+++ rpcs3/Emu/RSX/NV47/HW/nv406e.cpp
@@ -1,6 +1,7 @@
 #include "stdafx.h"
 #include "nv406e.h"
 #include "common.h"
+#include "nv47_sync.hpp"
 
 #include "Emu/RSX/RSXThread.h"
