+++ rpcs3/Emu/RSX/NV47/HW/nv406e.h
@@ -1,3 +1,4 @@
+// NV47 Sync Objects
 #pragma once
 
 #include "context.h"
