+++ rpcs3/Emu/RSX/NV47/HW/nv47.h
@@ -1,0 +1,8 @@
+// RSX Hardware definitions
+#pragma once
+
+#include "nv0039.h" // Buffer objects
+#include "nv3089.h" // Blit engine (2D)
+#include "nv308a.h" // Format conversion
+#include "nv406e.h" // Sync objects
+#include "nv4097.h" // 3D engine (GRAPH)
