+++ rpcs3/emucore.vcxproj
@@ -95,11 +95,13 @@
     <ClCompile Include="Emu\perf_monitor.cpp" />
     <ClCompile Include="Emu\RSX\Common\texture_cache.cpp" />
     <ClCompile Include="Emu\RSX\Core\RSXContext.cpp" />
-    <ClCompile Include="Emu\RSX\NV47\common.cpp" />
-    <ClCompile Include="Emu\RSX\NV47\nv3089.cpp" />
-    <ClCompile Include="Emu\RSX\NV47\nv308a.cpp" />
-    <ClCompile Include="Emu\RSX\NV47\nv406e.cpp" />
-    <ClCompile Include="Emu\RSX\NV47\nv4097.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\FW\reg_context.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\HW\common.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\HW\nv0039.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\HW\nv3089.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\HW\nv308a.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\HW\nv406e.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\HW\nv4097.cpp" />
     <ClCompile Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu.cpp" />
     <ClCompile Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu_components.cpp" />
     <ClCompile Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu_message_box.cpp" />
@@ -589,15 +591,20 @@
     <ClInclude Include="Emu\RSX\Core\RSXDisplay.h" />
     <ClInclude Include="Emu\RSX\Core\RSXReservationLock.hpp" />
     <ClInclude Include="Emu\RSX\Core\RSXVertexTypes.h" />
-    <ClInclude Include="Emu\RSX\NV47\context.h" />
-    <ClInclude Include="Emu\RSX\NV47\context_accessors.define.h" />
-    <ClInclude Include="Emu\RSX\NV47\context_accessors.undef.h" />
-    <ClInclude Include="Emu\RSX\NV47\nv3089.h" />
-    <ClInclude Include="Emu\RSX\NV47\nv308a.h" />
-    <ClInclude Include="Emu\RSX\NV47\nv406e.h" />
-    <ClInclude Include="Emu\RSX\NV47\nv4097.h" />
-    <ClInclude Include="Emu\RSX\NV47\nv47.h" />
-    <ClInclude Include="Emu\RSX\NV47\common.h" />
+    <ClInclude Include="Emu\RSX\NV47\FW\draw_call.hpp" />
+    <ClInclude Include="Emu\RSX\NV47\FW\draw_call.inc.h" />
+    <ClInclude Include="Emu\RSX\NV47\FW\reg_context.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\context.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\context_accessors.define.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\context_accessors.undef.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\nv0039.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\nv3089.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\nv308a.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\nv406e.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\nv4097.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\nv47.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\common.h" />
+    <ClInclude Include="Emu\RSX\NV47\HW\nv47_sync.hpp" />
     <ClInclude Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu.h" />
     <ClInclude Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu_components.h" />
     <ClInclude Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu_message_box.h" />
