+++ rpcs3/emucore.vcxproj.filters
@@ -100,6 +100,12 @@
     <Filter Include="Emu\GPU\RSX\NV47">
       <UniqueIdentifier>{213387bd-09c5-4247-8fb0-b3cae06ba34b}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Emu\GPU\RSX\NV47\HW">
+      <UniqueIdentifier>{24fc03a5-0469-4b6e-89df-44459e51855f}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Emu\GPU\RSX\NV47\FW">
+      <UniqueIdentifier>{fdb8ff8f-5a03-45a2-8c0a-16a89b7a574b}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Crypto\aes.cpp">
@@ -1213,20 +1219,26 @@
     <ClCompile Include="Emu\RSX\Program\SPIRVCommon.cpp">
       <Filter>Emu\GPU\RSX\Program</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\NV47\nv4097.cpp">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClCompile Include="Emu\RSX\NV47\HW\common.cpp">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\NV47\common.cpp">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClCompile Include="Emu\RSX\NV47\HW\nv0039.cpp">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\NV47\nv406e.cpp">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClCompile Include="Emu\RSX\NV47\HW\nv308a.cpp">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\NV47\nv308a.cpp">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClCompile Include="Emu\RSX\NV47\HW\nv406e.cpp">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\NV47\nv3089.cpp">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClCompile Include="Emu\RSX\NV47\HW\nv3089.cpp">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
+    </ClCompile>
+    <ClCompile Include="Emu\RSX\NV47\HW\nv4097.cpp">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
+    </ClCompile>
+    <ClCompile Include="Emu\RSX\NV47\FW\reg_context.cpp">
+      <Filter>Emu\GPU\RSX\NV47\FW</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -2467,32 +2479,47 @@
     <ClInclude Include="Emu\RSX\Program\SPIRVCommon.h">
       <Filter>Emu\GPU\RSX\Program</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\nv47.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\common.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\nv4097.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\context.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\nv406e.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\context_accessors.define.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\nv3089.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\context_accessors.undef.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\nv308a.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\nv0039.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\context.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\nv47.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\common.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\nv47_sync.hpp">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\context_accessors.define.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\nv308a.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\NV47\context_accessors.undef.h">
-      <Filter>Emu\GPU\RSX\NV47</Filter>
+    <ClInclude Include="Emu\RSX\NV47\HW\nv406e.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\NV47\HW\nv3089.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\NV47\HW\nv4097.h">
+      <Filter>Emu\GPU\RSX\NV47\HW</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\NV47\FW\draw_call.inc.h">
+      <Filter>Emu\GPU\RSX\NV47\FW</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\NV47\FW\reg_context.h">
+      <Filter>Emu\GPU\RSX\NV47\FW</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\NV47\FW\draw_call.hpp">
+      <Filter>Emu\GPU\RSX\NV47\FW</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
